occupancy-service/pyproject.toml
[project]
name = "occupancy-service"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "aio-pika>=9.5.7",
    "alembic>=1.17.0",
    "psycopg2-binary>=2.9.11",
    "fastapi[standard]>=0.118.0",
    "paho-mqtt>=2.1.0",
    "ruff>=0.13.2",
    "sqlmodel>=0.0.27",
    "pydantic-settings>=2.0.0",
    "python-dotenv>=1.2.1",
]

[dependency-groups]
dev = [
    "pre-commit>=4.3.0",
    "pytest>=8.4.2",
    "pytest-asyncio>=1.2.0",
    "pytest-cov>=7.0.0",
    "ruff>=0.14.1",
    "testcontainers[postgres,rabbitmq]>=4.13.2",
]

[tool.ruff]
# Exclude a variety of commonly ignored directories.
exclude = [
    ".bzr",
    ".direnv",
    ".eggs",
    ".git",
    ".git-rewrite",
    ".hg",
    ".ipynb_checkpoints",
    ".mypy_cache",
    ".nox",
    ".pants.d",
    ".pyenv",
    ".pytest_cache",
    ".pytype",
    ".ruff_cache",
    ".svn",
    ".tox",
    ".venv",
    ".vscode",
    "__pypackages__",
    "_build",
    "buck-out",
    "build",
    "dist",
    "node_modules",
    "site-packages",
    "venv",
    "migrations",
]

# Same as Black.
line-length = 88
indent-width = 4
# Python 3.13 compatibility
target-version = "py313"


[tool.ruff.lint]
select = [
    "E", # pycodestyle errors
    "F", # pyflakes errors
    "W", # pycodestyle warnings
    "I", # isort issues
    "TC", # test code issues
    "N", # naming convention issues
    "D", # pydocstyle issues
    "COM", # flake8-commas issues
    "C4", # flake8-comprehensions issues
    "PL", # pylint issues
    "PT", # pylint type issues
    "G", # flake8-annotations issues
    "ICN", # flake8-import-conventions issues
    "A", # flake8-annotations-complexity issues
    "B", # flake8-bugbear issues
    "S", # bandit security issues
    "ASYNC", # flake8-async issues
    "ANN", # flake8-annotations issues
    "FAST", # flake8-fastapi issues
    "SIM", # flake8-simplify issues
]
ignore = [
    "D100",  # Missing docstring in public module
    "D104",  # Missing docstring in public package
    "D107",  # Missing docstring in __init__
    "D203",  # 1 blank line required before class docstring
    "D213",  # Multi-line summary second line issue
    "COM812",  # Missing trailing comma before closing bracket
    "B008",  # Do not perform function calls in argument defaults
    "S101",  # Use of assert detected
]
# Allow fix for all enabled rules (when `--fix`) is provided.
fixable = ["ALL"]
unfixable = []
# Allow unused variables when underscore-prefixed.
dummy-variable-rgx = "^(_+|(_+[a-zA-Z0-9_]*[a-zA-Z0-9]+?))$"

[tool.ruff.format]
# Like Black, use double quotes for strings.
quote-style = "double"

# Like Black, indent with spaces, rather than tabs.
indent-style = "space"

# Like Black, respect magic trailing commas.
skip-magic-trailing-comma = false

# Like Black, automatically detect the appropriate line ending.
line-ending = "auto"
